+++ test/conftest.py
@@ -23,6 +23,7 @@ _PRISTINE_GLOBALS = {
     (mb_search, "search_releases"): mb_search.search_releases,
     (cover_art, "ensure_cover"): cover_art.ensure_cover,
     (cover_art, "check_front"): cover_art.check_front,
+    (cover_art, "fetch_image"): cover_art.fetch_image,
 }